Project-SF/requirements.txt
paho-mqtt>=1.6.1
influxdb-client>=1.36.0
fastapi>=0.109.0
uvicorn>=0.27.0
pandas>=2.2.0
python-dotenv>=1.0.0
requests>=2.31.0
scikit-learn>=1.3.0
